Moved DefaultHostValidator.ts to types/DefaultHostValidator.ts
