+++ src/DefaultHostValidator.ts
@@ -1,5 +1,5 @@
 import Host from './types/Host';
-import HostValidator from './HostValidator'
+import HostValidator from './types/HostValidator'
 
 let validate: HostValidator;
 validate = function (host: Host): Promise<Host> { return Promise.resolve(host) }
