+++ src/types/HostValidator.ts
@@ -1,4 +1,4 @@
-import Host from './types/Host';
+import Host from './Host';
 
 export default interface HostValidator {
     (host: Host): Promise<Host>
